# Avoid allocating FormFeature in HttpRequest.HasFormContentType