--- a/src/Shared/HttpExtensions.cs
+++ b/src/Shared/HttpExtensions.cs
@@ -1,30 +1,25 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
+using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Http;
 
 internal static class HttpExtensions
 {
-    internal const string UrlEncodedFormContentType = "application/x-www-form-urlencoded";
-    internal const string MultipartFormContentType = "multipart/form-data";
+    private const string UrlEncodedFormContentType = "application/x-www-form-urlencoded";
+    private const string MultipartFormContentType = "multipart/form-data";
 
     internal static bool IsValidHttpMethodForForm(string method) =>
         HttpMethods.IsPost(method) || HttpMethods.IsPut(method) || HttpMethods.IsPatch(method);
 
-    internal static bool IsValidContentTypeForForm(string? contentType)
+    internal static bool HasApplicationFormContentType([NotNullWhen(true)] string? contentType)
     {
-        if (contentType == null)
-        {
-            return false;
-        }
+        // Content-Type: application/x-www-form-urlencoded; charset=utf-8
+        return contentType != null && contentType.Contains(UrlEncodedFormContentType, StringComparison.OrdinalIgnoreCase);
+    }
 
-        // Abort early if this doesn't look like it could be a form-related content-type
-
-        if (contentType.Length < MultipartFormContentType.Length)
-        {
-            return false;
-        }
-
-        return contentType.Equals(UrlEncodedFormContentType, StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith(MultipartFormContentType, StringComparison.OrdinalIgnoreCase);
+    internal static bool HasMultipartFormContentType([NotNullWhen(true)] string? contentType)
+    {
+        // Content-Type: multipart/form-data; boundary=----WebKitFormBoundarymx2fSWqWSd0OxQqq
+        return contentType != null && contentType.Contains(MultipartFormContentType, StringComparison.OrdinalIgnoreCase);
     }
 }