Closes https://github.com/dotnet/aspnetcore/issues/50153.

```csharp
public class FormFeatureBenchmark
{
    private HttpRequest _request;

    [IterationSetup]
    public void Setup()
    {
        _request = new DefaultHttpRequest(new DefaultHttpContext());
    }

    [Benchmark]
    public void Get_HttpRequestHasFormContentType()
    {
        _ = _request.HasFormContentType;
    }
}
```

**Baseline**

|                            Method |     Mean |    Error |   StdDev |   Median |     Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------------- |---------:|---------:|---------:|---------:|---------:|------:|------:|------:|----------:|
| Get_HttpRequestHasFormContentType | 10.69 us | 1.879 us | 5.390 us | 9.208 us | 93,577.6 |     - |     - |     - |     776 B |

**With Changeset**

|                            Method |     Mean |    Error |   StdDev |      Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------------- |---------:|---------:|---------:|----------:|------:|------:|------:|----------:|
| Get_HttpRequestHasFormContentType | 8.959 us | 2.259 us | 6.445 us | 111,616.9 |     - |     - |     - |     672 B |